+++ Sprout/SproutCustomLogFormatter.m
@@ -2,17 +2,13 @@
 //  SproutCustomLogFormatter.m
 //
 //  Created by Levi Brown on October 4, 2012.
-//  Copyright (c) 2012-2016 Levi Brown <mailto:levigroker@gmail.com>
-//  This work is licensed under the Creative Commons Attribution 3.0
-//  Unported License. To view a copy of this license, visit
-//  http://creativecommons.org/licenses/by/3.0/ or send a letter to Creative
-//  Commons, 444 Castro Street, Suite 900, Mountain View, California, 94041,
-//  USA.
+//  Copyright (c) 2012-2017 Levi Brown <mailto:levigroker@gmail.com> This work is
+//  licensed under the Creative Commons Attribution 4.0 International License. To
+//  view a copy of this license, visit https://creativecommons.org/licenses/by/4.0/
+//  or send a letter to Creative Commons, PO Box 1866, Mountain View, CA 94042, USA.
 //
-//  The above attribution and the included license must accompany any version
-//  of the source code. Visible attribution in any binary distributable
-//  including this work (or derivatives) is not required, but would be
-//  appreciated.
+//  The above attribution and the included license must accompany any version of
+//  the source code, binary distributable, or derivatives.
 //
 
 #import "SproutCustomLogFormatter.h"
@@ -33,24 +29,21 @@
 - (NSString *)formatLogMessage:(DDLogMessage *)logMessage
 {
     NSString *logLevel = nil;
-    switch (logMessage->logFlag)
+    switch (logMessage->_flag)
     {
-        case LOG_FLAG_ERROR : logLevel = @"[ERROR]"; break;
-        case LOG_FLAG_WARN  : logLevel = @" [WARN]"; break;
-        case LOG_FLAG_INFO  : logLevel = @" [INFO]"; break;
-        default             : logLevel = @"[DEBUG]"; break;
+		case DDLogFlagError   : logLevel = @"[ERROR]"; break;
+		case DDLogFlagWarning : logLevel = @" [WARN]"; break;
+		case DDLogFlagInfo    : logLevel = @" [INFO]"; break;
+		case DDLogFlagDebug   :
+		default               : logLevel = @"[DEBUG]"; break;
     }
-    NSString *file = [NSString stringWithCString:logMessage->file encoding:NSASCIIStringEncoding];
-    file = [file lastPathComponent];
-    NSString *function = [NSString stringWithCString:logMessage->function encoding:NSASCIIStringEncoding];
-    NSString *timestamp = [self.dateFormatter stringFromDate:(logMessage->timestamp)];
 
-    char tidCStr[9];
-    int tidLen = snprintf(tidCStr, 9, "%x", logMessage->machThreadID);
-    NSString *threadID = [NSString stringWithCString:tidCStr encoding:NSASCIIStringEncoding];
-    threadID = [threadID substringToIndex:MIN((size_t)8, tidLen)];
+	NSString *file = [logMessage->_file lastPathComponent];
+    NSString *function = logMessage->_function;
+    NSString *timestamp = [self.dateFormatter stringFromDate:(logMessage->_timestamp)];
+    NSString *threadID = logMessage->_threadID;
 
-	return [NSString stringWithFormat:@"%@         <%@> %@(%@ %d)\n%@ %@ %@", timestamp, threadID, function, file, logMessage->lineNumber, timestamp, logLevel, logMessage->logMsg];
+	return [NSString stringWithFormat:@"%@         <%@> %@(%@ %d)\n%@ %@ %@", timestamp, threadID, function, file, (int)logMessage->_line, timestamp, logLevel, logMessage->_message];
 }
 
 @end