+++ SproutLib/Pods/Headers/Private/Sprout/CrashlyticsLogger.h
@@ -1,1 +1,0 @@
-../../../../../Sprout/CrashlyticsLogger.h